refactor: prompt-evals を tools/opencode-probe へ移動する
OpenCode 専用の測定器具（合成プロンプト捕捉プローブ＋決定的スモークゲート）であり、
promptfoo ベースの eval/ と紛らわしい名前とリポジトリ直下の配置を改める。

- git mv による履歴保持。実装ロジックは不変（process-tree.mjs は同一内容）
- tsconfig.prompt-evals.json → tsconfig.opencode-probe.json
- npm script は test:opencode-probe を追加し、test:prompt-evals を互換エイリアスとして維持
- .coderabbit.yaml の path_filters へ除外を追加（測定器具でプロダクション経路に乗らないため）
- CI・test-classification・docs の参照を全数書き換え（旧パス参照0件を機械確認）
