+++ .coderabbit.yaml
@@ -36,6 +36,8 @@ reviews:
     # 修正提案の対象にしない（直すと測定器が壊れる）。
     - "!eval/fixtures/**"
     - "!eval/cases/**"
+    # OpenCode の測定器具。プロダクション経路に乗らない。
+    - "!tools/opencode-probe/**"
     - "!dist/**"
     - "!node_modules/**"
     - "!package-lock.json"
