+++ tools/opencode-probe/fixtures/run-smoke-batch.mjs
