+++ tools/opencode-probe/fixtures/smoke-case.mjs
