+++ tools/opencode-probe/opencode-probe-lifecycle.d.mts
