+++ tools/opencode-probe/opencode-probe-lifecycle.mjs
