+++ tools/opencode-probe/plugin-probe.mjs
