+++ tools/opencode-probe/probe-entrypoint.d.mts
