+++ tools/opencode-probe/probe-entrypoint.mjs
