+++ tools/opencode-probe/probe-environment.d.mts
