+++ tools/opencode-probe/probe-environment.mjs
