+++ tools/opencode-probe/probe-process.d.mts
