+++ tools/opencode-probe/probe-process.mjs
