+++ tools/opencode-probe/probe-workspace.d.mts
