+++ tools/opencode-probe/probe-workspace.mjs
