+++ tools/opencode-probe/process-tree.mjs
