+++ tools/opencode-probe/prompts/opencode-base-1.17.16.md
