+++ tools/opencode-probe/prompts/round1.txt
