+++ tools/opencode-probe/prompts/round2-base.txt
