+++ tools/opencode-probe/prompts/round2-overlay.txt
