+++ tools/opencode-probe/sdk-prompt-capture.mjs
