+++ tools/opencode-probe/sdk-tool-eval.mjs
