+++ tools/opencode-probe/smoke-process.d.mts
