+++ tools/opencode-probe/smoke-process.mjs
