+++ tools/opencode-probe/summarize-probe.mjs
